pyproject.toml
[tool.poetry]
name = "pythoneda-artifact-application-pyproject-versioning"
version = "0.0.1a2"
description = "Application layer for pythoneda-artifact/pyproject-versioning"
authors = ["rydnr <github@acm-sl.org>"]
readme = "README.md"
homepage = "https://github.com/pythoneda-artifact-application/pyproject-versioning"
packages = [{include = "pythonedaartifactapplicationpyprojectversioning"}]

[tool.poetry.dependencies]
dbus-next = "^0.2.3"
grpcio = "^1.41.0"
python = "^3.8"
pythoneda-application-base = "^0.0.1a12"
pythoneda-artifact-event-versioning = "^0.0.1a1"
pythoneda-artifact-event-infrastructure-versioning = "^0.0.1a1"
pythoneda-artifact-pyproject-versioning = "^0.0.1a1"
pythoneda-artifact-infrastructure-pyproject-versioning = "^0.0.1a1"
pythoneda-base = "^0.0.1a16"
pythoneda-infrastructure-base = "^0.0.1a12"
pythoneda-shared-git = "^0.0.1a3"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
